# stage2/docs/backend_openapi.json
{"openapi":"3.1.0","info":{"title":"Task Manager API","description":"API для управления задачами с аутентификацией","version":"1.0.0"},"paths":{"/register":{"post":{"summary":"Регистрация нового пользователя","description":"Создает нового пользователя в системе. Требует email и пароль.","operationId":"register_user_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Пользователь успешно создан","content":{"application/json":{"schema":{"type":"object","title":"Response Register User Register Post"},"example":{"message":"User created successfully"}}}},"400":{"description":"Email уже зарегистрирован","content":{"application/json":{"example":{"detail":"Email already registered"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/token":{"post":{"summary":"Получение токена доступа","description":"Аутентифицирует пользователя и возвращает JWT токен для доступа к API.","operationId":"login_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_token_post"}}},"required":true},"responses":{"200":{"description":"Успешная аутентификация","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"},"example":{"access_token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...","token_type":"bearer"}}}},"401":{"description":"Неверные учетные данные","content":{"application/json":{"example":{"detail":"Incorrect username or password"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/token/refresh":{"post":{"summary":"Refresh Token","operationId":"refresh_token_token_refresh_post","parameters":[{"name":"refresh_token","in":"query","required":true,"schema":{"type":"string","title":"Refresh Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/token/revoke":{"post":{"summary":"Revoke Token","operationId":"revoke_token_token_revoke_post","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/":{"get":{"summary":"Получение списка задач","description":"Возвращает список всех задач пользователя. Требует JWT токен.","operationId":"read_tasks_tasks__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}},{"name":"is_completed","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Completed"}},{"name":"sort_by","in":"query","required":false,"schema":{"type":"string","description":"Поле для сортировки (created_at или is_completed)","default":"created_at","title":"Sort By"},"description":"Поле для сортировки (created_at или is_completed)"},{"name":"sort_order","in":"query","required":false,"schema":{"type":"string","description":"Порядок сортировки (asc или desc)","default":"desc","title":"Sort Order"},"description":"Порядок сортировки (asc или desc)"}],"responses":{"200":{"description":"Список задач успешно получен","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"},"title":"Response Read Tasks Tasks  Get"},"example":[{"id":1,"title":"Задача 1","description":"Описание задачи","is_completed":false,"created_at":"2024-01-01T12:00:00","owner_id":1}]}}},"401":{"description":"Неверный токен доступа","content":{"application/json":{"example":{"detail":"Invalid authentication credentials"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Создание новой задачи","description":"Создает новую задачу для текущего пользователя. Требует JWT токен.","operationId":"create_task_tasks__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCreate"}}}},"responses":{"200":{"description":"Задача успешно создана","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"},"example":{"id":1,"title":"Новая задача","description":"Описание задачи","is_completed":false,"created_at":"2024-01-01T12:00:00","owner_id":1}}}},"401":{"description":"Неверный токен доступа","content":{"application/json":{"example":{"detail":"Invalid authentication credentials"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/{task_id}":{"put":{"summary":"Обновление задачи","description":"Обновляет существующую задачу. Требует JWT токен.","operationId":"update_task_tasks__task_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskUpdate"}}}},"responses":{"200":{"description":"Задача успешно обновлена","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"},"example":{"id":1,"title":"Обновленная задача","description":"Новое описание","is_completed":true,"created_at":"2024-01-01T12:00:00","owner_id":1}}}},"401":{"description":"Неверный токен доступа","content":{"application/json":{"example":{"detail":"Invalid authentication credentials"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Удаление задачи","description":"Удаляет существующую задачу. Требует JWT токен.","operationId":"delete_task_tasks__task_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"integer","title":"Task Id"}}],"responses":{"200":{"description":"Задача успешно удалена","content":{"application/json":{"schema":{"type":"object","title":"Response Delete Task Tasks  Task Id  Delete"},"example":{"message":"Task deleted successfully"}}}},"401":{"description":"Неверный токен доступа","content":{"application/json":{"example":{"detail":"Invalid authentication credentials"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_token_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Task":{"properties":{"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"is_completed":{"type":"boolean","title":"Is Completed","default":false},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"owner_id":{"type":"integer","title":"Owner Id"}},"type":"object","required":["title","id","created_at","owner_id"],"title":"Task"},"TaskCreate":{"properties":{"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"is_completed":{"type":"boolean","title":"Is Completed","default":false}},"type":"object","required":["title"],"title":"TaskCreate"},"TaskUpdate":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"is_completed":{"type":"boolean","title":"Is Completed","default":false}},"type":"object","title":"TaskUpdate"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"UserCreate":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}